src/presentation/controllers/appointment/load-appointment-by-id/load-appointment-by-id-controller.spec.ts
import { LoadAppointmentByIdController } from './load-appointment-by-id-controller'
import { HttpRequest, LoadAppointmentById } from './load-appointment-by-id-controller-protocols'
import MockDate from 'mockdate'
import { forbidden, ok, serverError } from '@/presentation/helpers/http/http-helper'
import { mockAppointmentModel, throwError } from '@/domain/test'
import { mockLoadAppointmentById } from '@/presentation/test'
import { InvalidParamError } from '@/presentation/errors'

const mockRequest = (): HttpRequest => ({
  params: {
    id: 'any_id'
  }
})

type SutTypes = {
  sut: LoadAppointmentByIdController
  loadAppointmentByIdStub: LoadAppointmentById
}

const makeSut = (): SutTypes => {
  const loadAppointmentByIdStub = mockLoadAppointmentById()
  const sut = new LoadAppointmentByIdController(loadAppointmentByIdStub)
  return {
    sut,
    loadAppointmentByIdStub
  }
}

describe('DeleteAppointment Controller', () => {
  beforeAll(() => {
    MockDate.set(new Date())
  })

  afterAll(() => {
    MockDate.reset()
  })

  test('Should call LoadAppointmentById with correct values', async () => {
    const { sut, loadAppointmentByIdStub } = makeSut()
    const loadByIdSpy = jest.spyOn(loadAppointmentByIdStub, 'loadById')
    await sut.handle(mockRequest())
    expect(loadByIdSpy).toHaveBeenCalledWith('any_id')
  })

  test('Should return 403 if LoadAppointmentById returns null', async () => {
    const { sut, loadAppointmentByIdStub } = makeSut()
    jest.spyOn(loadAppointmentByIdStub, 'loadById').mockReturnValueOnce(Promise.resolve(null as any))
    const httpResponse = await sut.handle(mockRequest())
    expect(httpResponse).toEqual(forbidden(new InvalidParamError('id')))
  })

  test('Should return 500 if LoadAppointmentById throws', async () => {
    const { sut, loadAppointmentByIdStub } = makeSut()
    jest.spyOn(loadAppointmentByIdStub, 'loadById').mockImplementationOnce(throwError)
    const httpResponse = await sut.handle(mockRequest())
    expect(httpResponse).toEqual(serverError(new Error()))
  })

  test('Should return 200 if valid data is provided', async () => {
    const { sut } = makeSut()
    const httpResponse = await sut.handle(mockRequest())
    expect(httpResponse).toEqual(ok(mockAppointmentModel()))
  })
})
